Fix race condition on ScrutinyClient connect (#213)
Fixed race condition that could hang the client for 2 seconds after connecting the ScrutinyClient. The server status was received before we started to wait for the first update. Was waiting for the 2nd update.   Now doesn't wait if already received.

This problem was what has been slowing down the unit tests years now. (yay!)